fix(rescue-1-audit): apply 6 audit findings from Rescue-1 mid-phase audit
External red-team audit (sub-agent a47de986) reviewed commits 018dd6e + c132eb8.
Verdict: Ready-with-fixes, 0 BLOCKER, 83/100 confidence к Premium phase.
6 of 9 findings applied (3 deferred к backlog as cosmetic / future-proofing).

Applied:

  * PI-RESCUE-02 MEDIUM: synthetic_posterior intercept dtype mismatch
    - Cast intercept_samples к np.float32 to match bayesian_engine real output
    - Closes schema contract gap between synthetic + real PyMC outputs

  * PI-RESCUE-05 LOW: ModeBadge tooltip Escape dismiss
    - Added handleKeydown for Escape key
    - Mounted via svelte:window for global capture when tooltip open
    - Accessibility improvement for keyboard-only users

  * PI-RESCUE-06 LOW: generate_pymc_fixture dead code branch removed
    - if/else assigned identically — replaced с single arr = stacked.values

  * PI-RESCUE-07 MEDIUM: bias_check NaN bypass
    - Added math.isfinite(predicted_mean) check before ε threshold
    - Previously: abs(NaN) < 1e-6 evaluated False → division by NaN
    - Now: explicit NaN diagnostic warning emitted

  * PI-RESCUE-08 LOW: ProcessLock blocking=True error masking
    - Both _acquire_windows + _acquire_posix raise on OSError regardless of blocking
    - Previously blocking=True silently returned False, masking permission/ENFILE errors

  * PI-RESCUE-09 LOW: AutosaveManager double signal registration
    - Added _signal_handlers_installed flag guard
    - Idempotent _install_signal_handlers — second call no-op

Deferred (cosmetic / future-proofing):
  * PI-RESCUE-01 LOW: is_pickle_magic protocol 0/1 detection comment-reality mismatch
    (no security hole — pickle fallback path doesn't exist)
  * PI-RESCUE-03 LOW: GitHub actions first-party @v4 floating tag
    (supply-chain hardening backlog; community actions SHA-pinned)
  * PI-RESCUE-04 LOW: navigator.platform deprecated warning
    (still works в all current browsers; v0.2.0 migration)

Tests: 123 pass in audit-touched suites (phase Π.2 + Σ.0.4 + Rescue-1.5).
Full suite verification next.

Self-audit pattern: 9 findings caught by external sub-agent, 6 real applied,
3 deferred. Sub-agent quality высокая — нашёл NaN bypass и dtype mismatch что
я не заметила во self-review.

+++ src/aurora_launch/sample_bundles/synthetic_posterior.py
@@ -301,13 +301,15 @@ def derive_synthetic_posterior(
 
     # R-12 audit fix: intercept + control_betas keys present in posterior
     # (decomposer expects these from real PyMC output; absence → KeyError).
+    # PI-RESCUE-02 audit fix: cast intercept к float32 to match bayesian_engine
+    # real output (line 903 of bayesian_engine.py).
     intercept_samples = _generate_samples_normal(
         mean=0.0,  # normalised intercept around zero baseline
         std=DEFAULT_INTERCEPT_STD,
         n_samples=n_samples,
         clip_negative=False,  # intercept can be negative in normalised space
         rng=rng,
-    )
+    ).astype(np.float32)
     # No control variables в synthetic case — shape (0, n_samples) empty array
     control_betas_samples = np.zeros((0, n_samples), dtype=np.float32)
